+++ GenerateFilters/GenerateFilters.csproj
@@ -17,10 +17,10 @@
     <SignAssembly>true</SignAssembly>
   </PropertyGroup>
   <PropertyGroup>
-    <AssemblyOriginatorKeyFile>Key.snk</AssemblyOriginatorKeyFile>
+    <AssemblyOriginatorKeyFile>key.snk</AssemblyOriginatorKeyFile>
   </PropertyGroup>
   <PropertyGroup>
-    <ApplicationIcon>Resources\GenerateFilters.ico</ApplicationIcon>
+    <ApplicationIcon>resources\GenerateFilters.ico</ApplicationIcon>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x64'">
     <DebugSymbols>true</DebugSymbols>
@@ -77,15 +77,19 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="GenerateFilters.cs" />
-    <Compile Include="GenerateFiltersPackage.cs" />
+    <Compile Include="source\filter\filter.cs" />
+    <Compile Include="source\filter\work.cs" />
+    <Compile Include="source\package.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="source\filter\path.cs" />
+    <Compile Include="source\filter\project.cs" />
+    <Compile Include="source\filter\xml.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="Resources\VCProjectEngine_17.0.dll" />
-    <None Include="Key.snk" />
-    <Content Include="Resources\GenerateFiltersPackage.ico" />
-    <Content Include="Resources\LICENSE">
+    <Content Include="resources\VCProjectEngine_17.0.dll" />
+    <None Include="key.snk" />
+    <Content Include="resources\GenerateFiltersPackage.ico" />
+    <Content Include="resources\LICENSE">
       <IncludeInVSIX>true</IncludeInVSIX>
     </Content>
     <None Include="source.extension.vsixmanifest">
@@ -93,20 +97,20 @@
     </None>
   </ItemGroup>
   <ItemGroup>
-    <VSCTCompile Include="GenerateFiltersPackage.vsct">
+    <VSCTCompile Include="packageXml.vsct">
       <ResourceName>Menus.ctmenu</ResourceName>
       <SubType>Designer</SubType>
     </VSCTCompile>
-    <Content Include="Resources\GenerateFilters.ico">
+    <Content Include="resources\GenerateFilters.ico">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
       <IncludeInVSIX>true</IncludeInVSIX>
     </Content>
-    <Content Include="Resources\PreviewImage.png">
+    <Content Include="resources\PreviewImage.png">
       <IncludeInVSIX>true</IncludeInVSIX>
     </Content>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="VSPackage.resx">
+    <EmbeddedResource Include="resources.resx">
       <MergeWithCTO>true</MergeWithCTO>
       <ManifestResourceName>VSPackage</ManifestResourceName>
       <SubType>Designer</SubType>
@@ -115,6 +119,11 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Xml" />
+    <Reference Include="VCProjectEngine_17.0, Version=17.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>resources\VCProjectEngine_17.0.dll</HintPath>
+      <EmbedInteropTypes>True</EmbedInteropTypes>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.VisualStudio.SDK">
