+++ GenerateFilters/packageXml.vsct
