+++ GenerateFilters/source/other.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.ComponentModel.Design;
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
+using Microsoft.Internal.VisualStudio.PlatformUI;
+using EnvDTE;
+using System.Collections.Generic;
+using System.IO;
+using System.Xml;
+using System.Reflection;
+using System.Text;
+
+namespace VisualStudioCppExtensions
+{
+    internal sealed partial class GenerateFilter
+    {
+
+
+
+
+
+
+
+
+    }
+}
