+++ GenerateFilters/source/package.cs
@@ -1,0 +1,42 @@
+﻿using Microsoft.VisualStudio;
+using Microsoft.VisualStudio.Shell;
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.InteropServices;
+using System.Threading;
+using System.Threading.Tasks;
+
+
+
+
+namespace VisualStudioCppExtensions
+{
+
+
+
+    [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
+    [InstalledProductRegistration("#110", "#112", "1.0", IconResourceID = 400)]
+    [ProvideMenuResource("Menus.ctmenu", 1)]
+    [Guid(PackageGuidString)]
+    [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1650:ElementDocumentationMustBeSpelledCorrectly", Justification = "pkgdef, VS and vsixmanifest are valid VS terms")]
+    [ProvideAutoLoad(VSConstants.UICONTEXT.SolutionExistsAndFullyLoaded_string, PackageAutoLoadFlags.BackgroundLoad)]
+    public sealed class package2 : AsyncPackage
+    {
+        public const string PackageGuidString = "99d03761-6200-41ad-b2a1-638ae9e780e5";
+
+        public package2() { }
+
+
+
+        protected override async Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
+        {
+            filter.Initialize(this);
+            await JoinableTaskFactory.SwitchToMainThreadAsync(cancellationToken);
+        }
+    }
+
+
+
+
+
+}
